Initial state of project
Goal of this project : Capture live data from Gamecube-to-USB adapters,
and parse it to a human-readable format (CSV file).

To do this, the application reads data pipe "/dev/usbmonX" (X =
appropriate USB bus number), special pipes available on Linux for
listening to USB traffic. As such, this application is only supported
for Linux, although eventually supporting other platforms would be nice.

Furthermore, this application uses multi-threading which is (basically)
impossible in CPython, the native Python interpreter. So instead, this
application should be ran with PyPy interpreter.

Project is comprised of 3 source files so far :
- core.py : All the logic for capturing raw binary stream, parsing its
data, and writing it to an output file.
- gc_conversion.py : Defines objects that can parse the data of captured
USB packets to infer the state of the controller (A button pressed?
C-Stick tilted up?)
- user_interface.py : Handles user interaction through CLI. 2
configuration modes exist for now, with an interactive prompt (powered
by "questionary"), or with flags (powered by "click").