build: Reorganize project layout
Having script files flat in project root directory was cumbersome. While
the project was still simple enough to configure, this might not be the
case if I start splitting up code in more files, and storing them in
various folders. Which is likely to happen soon as I'll tackle design
review.
So it's better to be safe with a standard Python project layout, with
"src/(package name)". A new empty "__init__.py" file has been added, and
project imports have been adjusted to take the new project structure
into account.