chore!: License project under GPLv3
The intent behind this license choice is to make sure that modifications
and redistributions are readily available to gamer communities using
this tool in the hope that improvements are readily available to
everyone, rather than locked behind proprietary offers, so future
improvements will equally benefit everyone. Broadly speaking, every
player should be able to enjoy playing the game they like the way they
want with the supporting tools of their choice. So this software is and
will remain FOSS.